requirement.txt
appdirs==1.4.4
appnope==0.1.2
asgiref==3.4.1
astroid==2.5.6
autopep8==1.5.5
backcall==0.2.0
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorgram.py==1.2.0
decorator==5.0.8
distlib==0.3.2
dominate==2.6.0
fastapi==0.66.1
filelock==3.0.12
flake8==3.8.4
Flask==2.0.1
Flask-Bootstrap==3.3.7.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
fpdf==1.7.2
greenlet==1.1.0
h11==0.12.0
idna==2.10
ipykernel==5.5.5
ipython==7.23.1
ipython-genutils==0.2.0
isort==5.8.0
itsdangerous==2.0.1
jedi==0.17.2
Jinja2==3.0.1
jupyter-client==6.1.12
jupyter-core==4.7.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mccabe==0.6.1
mypy==0.812
mypy-extensions==0.4.3
numpy==1.20.3
packaging @ file:///private/tmp/sip--packaging-20210512-78007-1o5u9nj/packaging-20.9
pandas==1.2.4
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
platformdirs==2.0.2
pluggy==0.13.1
prompt-toolkit==3.0.18
psycopg2==2.9.1
psycopg2-binary==2.9.1
ptyprocess==0.7.0
pycodestyle==2.6.0
pydantic==1.8.2
pydocstyle==6.1.1
pyflakes==2.2.0
Pygments==2.9.0
pylint==2.8.3
pyls-mypy @ git+https://github.com/tomv564/pyls-mypy.git@3b105b491a42eec304f19e6ed6da1053d12958c6
pyparsing==2.4.7
pyperclip==1.8.2
PyQt3D==5.15.4
PyQt5==5.15.4
PyQt5-sip==12.9.0
PyQtChart==5.15.4
PyQtDataVisualization==5.15.4
PyQtNetworkAuth==5.15.4
PyQtPurchasing==5.15.4
PyQtWebEngine==5.15.4
python-dateutil==2.8.1
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
pytz==2021.1
pyzmq==22.0.3
QScintilla==2.12.1
requests==2.25.1
rope==0.19.0
selenium==3.141.0
sip==6.1.0
six==1.16.0
snowballstemmer==2.1.0
soupsieve==2.2.1
spotipy==2.18.0
SQLAlchemy==1.4.22
starlette==0.14.2
TBB==0.2
tk==0.1.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
typed-ast==1.4.3
typing-extensions==3.10.0.0
ujson==4.0.2
urllib3==1.26.5
uvicorn==0.14.0
virtualenv==20.6.0
visitor==0.1.3
wcwidth==0.2.5
Werkzeug==2.0.1
wrapt==1.12.1
WTForms==2.3.3
yapf==0.31.0
